Use size_t for size constants

+++ include/md5.hpp
@@ -102,10 +102,10 @@ namespace md5
         return rotateLeft(a + i(b, c, d) + x + ac, sh) + b;
     }
 
-    constexpr uint32_t digestIntCount = 4; // number of 32bit integers per MD5 digest
-    constexpr uint32_t digestByteCount = digestIntCount * 4;
-    constexpr uint32_t blockIntCount = 16; // number of 32bit integers per MD5 block
-    constexpr uint32_t blockByteCount = blockIntCount * 4;
+    constexpr size_t digestIntCount = 4; // number of 32bit integers per MD5 digest
+    constexpr size_t digestByteCount = digestIntCount * 4;
+    constexpr size_t blockIntCount = 16; // number of 32bit integers per MD5 block
+    constexpr size_t blockByteCount = blockIntCount * 4;
     using Block = uint8_t[blockByteCount];
     using State = uint32_t[digestIntCount];
