+++ include/sha1.hpp
@@ -12,10 +12,10 @@
 
 namespace sha1
 {
-    constexpr uint32_t digestIntCount = 5; // number of 32bit integers per SHA1 digest
-    constexpr uint32_t digestByteCount = digestIntCount * 4;
-    constexpr uint32_t blockIntCount = 16; // number of 32bit integers per SHA1 block
-    constexpr uint32_t blockByteCount = blockIntCount * 4;
+    constexpr size_t digestIntCount = 5; // number of 32bit integers per SHA1 digest
+    constexpr size_t digestByteCount = digestIntCount * 4;
+    constexpr size_t blockIntCount = 16; // number of 32bit integers per SHA1 block
+    constexpr size_t blockByteCount = blockIntCount * 4;
     using Block = uint8_t[blockByteCount];
     using State = uint32_t[digestIntCount];
